Adjust split_neuron instrumentation
