+++ rs/tla_instrumentation/tla_instrumentation/tests/structs.rs
@@ -1,18 +1,19 @@
 use std::{
     collections::{BTreeMap, BTreeSet},
-    path::PathBuf,
     ptr::addr_of_mut,
 };
 
 // Also possible to define a wrapper macro, in order to ensure that logging is only
 // done when certain crate features are enabled
 use tla_instrumentation::{
-    checker::{check_tla_code_link, PredicateDescription},
     tla_log_locals, tla_log_request, tla_log_response,
     tla_value::{TlaValue, ToTla},
     Destination, InstrumentationState,
 };
 use tla_instrumentation_proc_macros::tla_update_method;
+
+mod common;
+use common::check_tla_trace;
 
 // Example of how to separate as much of the instrumentation code as possible from the main code
 #[macro_use]
@@ -148,30 +149,6 @@ impl StructCanister {
     }
 }
 
-// Add JAVABASE/bin to PATH to make the Bazel-provided JRE available to scripts
-fn set_java_path() {
-    let current_path = std::env::var("PATH").unwrap();
-    let bazel_java = std::env::var("JAVABASE").unwrap();
-    std::env::set_var("PATH", format!("{current_path}:{bazel_java}/bin"));
-}
-
-/// Returns the path to the TLA module (e.g. `Foo.tla` -> `/home/me/tla/Foo.tla`).
-/// TLA modules are read from $TLA_MODULES (space-separated list)
-/// NOTE: this assumes unique basenames amongst the modules
-fn get_tla_module_path(module: &str) -> PathBuf {
-    let modules = std::env::var("TLA_MODULES").expect(
-        "environment variable 'TLA_MODULES' should be a space-separated list of TLA modules",
-    );
-
-    modules
-        .split(" ")
-        .map(|f| f.into()) /* str -> PathBuf */
-        .find(|f: &PathBuf| f.file_name().is_some_and(|file_name| file_name == module))
-        .unwrap_or_else(|| {
-            panic!("Could not find TLA module {module}, check 'TLA_MODULES' is set correctly")
-        })
-}
-
 #[test]
 fn size_test() {
     let myval = TlaValue::Record(BTreeMap::from([
@@ -305,33 +282,5 @@ fn struct_test() {
         Some(&Vec::<TlaValue>::new().to_tla_value())
     );
 
-    set_java_path();
-
-    let apalache = std::env::var("TLA_APALACHE_BIN")
-        .expect("environment variable 'TLA_APALACHE_BIN' should point to the apalache binary");
-    let apalache = PathBuf::from(apalache);
-
-    if !apalache.as_path().is_file() {
-        panic!("bad apalache bin from 'TLA_APALACHE_BIN': '{:?}'", apalache);
-    }
-
-    let update = trace.update.clone();
-    for pair in &trace.state_pairs {
-        let constants = trace.constants.clone();
-        println!("Constants: {:?}", constants);
-        // NOTE: the 'process_id" is actually the tla module name
-        let tla_module = format!("{}_Apalache.tla", update.process_id);
-        let tla_module = get_tla_module_path(&tla_module);
-        check_tla_code_link(
-            &apalache,
-            PredicateDescription {
-                tla_module,
-                transition_predicate: "Next".to_string(),
-                predicate_parameters: Vec::new(),
-            },
-            pair.clone(),
-            constants,
-        )
-        .expect("TLA link check failed");
-    }
+    check_tla_trace(trace);
 }
